Copy arguments before passing them to hypotheses
The problem is that data may be mutable, and previously Hypothesis
was just assuming that it would not be mutated. If this assumption
was falsified then it would *sometimes* be detected by Hypothesis'
s flakiness detection but not always, and even then it was not
entirely obvious what was going on because the data would have
been changed and thus so would the repr.

This commit introduces two concepts:

  1. A strategy has a has_immutable_data property which is True if
     all data reachable from a value produced by this strategy is
     immutable and thus you don't need to worry.
  2. A strategy also has a copy method. If has_immutable_data is
     True then copying just returns its value, otherwise it calls
     deepcopy. Note that deepcopy also has some optimisations for
     immutable data, but we have the advantage here that we can
     calculate it once without doing the O(n) cost incurred by
     deepcopy so in these cases it should be a lot faster.

This fixes https://github.com/DRMacIver/hypothesis/issues/19