Add internal assertions for could_have_produced
This was intended purely as a self-check I expected to always pass
but it actually uncovered two examples where I needed to have
implemented could_have_produced and didn't: JustStrategy and
StatefulStrategy. So this commit also implements those and expands
the testing around this to make sure I'm not missing any other examples.