Miscellaneous improvements around repr
This started by adding a test to the recursive tests which verified
that if you eval the nice_string for a descriptor you get the descriptor
back. Turns out you don't: There were some bugs around empty sets which
caused them to be represented as {} which of course is in fact a dict in
python and not a set (yay ambiguous notation).

The fallback strategy of using the __dict__ is a bit questionable too. I've
left it in for now but really this entire section of the code should be
reconsidered. However it was playing havoc with RandomWithSeed, so I ended
up special casing it out by having it respect the repr of any objects it
encounters via Just.

My first attempt to test this went via the JustStrategy which I then
realised deserved a better representation, so JustStrategy is now repred
as JustStrategy(value=...) rather than JustStrategy(Just(value=...)) which
was silly. So that fix is bundled in here too.

This also changes RandomWithSeed's repr to declare that it's actually a
RandomWithSeed rather than just a Random. That was mostly to placate the
test, but on reflection I think it makes more sense: It's not really great
to have the representation of an object be dishonest like that and will
probably lead to really surprising debugging events.