A prototype of a storage and serialization layer
This is at the moment entirely useless. It understands how to convert
all supported types to json-like objects which can then be marshalled
to and from json. It implements a prototype in memory storage layer
that will have *some* use but is mostly of interest for testing purposes
and will likely not be available in the final production version.

It does however contain what is probably pretty close to the final form
of the APIs. The serialization framework seems to work rather well and
the storage interface appears to be basically correct although may need
some tinkering with later.