Switch to an internal name for xrange
This is entirely to placate pylint which dislikes me shadowing xrange
but I feel like it has a point so I'm going to comply with its wishes.
Saying xrange but meaning our own custom definition of it is a bit
cheeky.

This also disables pylint on the compat module entirely. It will never
be happy with that file.