Add a notion of a format and basic data
This abstracts over the data storage format - the default remains
JSON but this would make it easy for someone who wants to e.g. use
msgpack to do so.

I'm not sure there's any real need to do so but it was easy to do
and seems consistent with Hypothesis's strategy of being really far
more flexible than is conceivably necessary but providing relatively
easy to use defaults so you don't have to touch it.