Factor out the example loop from Verifier
This currently has far simpler logic about how to choose the
parameters. However all the tests are passing with this simpler logic
so I might actually declare it good enough and only implement something
smarter when we have interesting example discovery working and can use
it to guide the search much more effectively.